reorg ast json
